Remove nix bits from repo and CI
These workflows fail often and I have no time or desire to learn
nix to troubleshoot and resolve them.

I have not been able to reach those that originally contributed
them by mentions on commits on GH, and since I hate to have
something that is essentially broken and opaque to me sitting
in the repo, with no maintainers, I plan to remove it.

See also my original position, and my later acceptance that
was hoping that the community would be more active in looking
after this:

* https://github.com/wez/wezterm/pull/3547#issuecomment-1537504636
* https://github.com/wez/wezterm/pull/3547#issuecomment-1831941228

If this is to remain, I think I will need to have someone commit
to looking after this stuff, and to commit to finding a replacement
in the event that they no longer have the capacity to maintain it.