Add missing docs, fix typo, bring to attention error. (#5150)
* Update swallow_mouse_click_on_window_focus.md

Fix typo.

* Update appearance.md

URL in docs is no longer value. As a placeholder I put in the file where color_schemes can be found. I'm thinking that's not the right fix, but there is now awareness of the broken URL.

* Update window_frame.md

Add font and font_size description for window_frame.

* Update window_frame.md

Fix typo