Fix notch avoidance padding in full screen mode
Fixes https://github.com/wez/wezterm/issues/3807

The values returned by [`safeAreaInsets`][0] need to be multipled by the
scale factor of the display (in case of the MacBook Pro 14-inch, 2021, a
value of 2.0).

There's a code comment referencing [#1737 (comment)][1], where an extra
`2` was added to this number, but something must have changed since then
because I cannot see a need for this now.

[0]: https://developer.apple.com/documentation/appkit/nsscreen/3882821-safeareainsets
[1]: https://github.com/wez/wezterm/issues/1737#issuecomment-1085923867