Handle # and ? characters in directory path
When referencing the current-working-directory, before
it is set by an OSC 7 escape sequence, we ask the OS
for the correct path.  This path was then being parsed
as a URL; where a "#" or "?" character would be
interpreted as the start of a fragment or query
component of a URL -- which is a mistake.

So this change parses the returned directory as such,
where those characters will be treated as a normal
character in the path.

Nothing is changed for the OSC 7 escape sequence case.
In that case, the application must percent-encode the
path before sending, so that those characters are not
misinterpreted.

As per issue #6158 reported by Syntaxheld