Copy terminfo into MacOS app resources
Currently the terminfo file has to be downloaded from the repo and
installed manually. It'd be nice to include it in the app so that a
package manager like Homebrew could install it in the right place when
Wezterm is installed, which is something that Alacritty does.

I've chosen to include the compiled version so that package manager
don't have to invoke `tic` and can just copy files around.

This will result in the following being added to the Resources folder:
```
WezTerm.app/Contents/Resources/terminfo/77/wezterm
```

Once this is in I'll send a patch to update the [Homebrew
formula](https://github.com/Homebrew/homebrew-cask/blob/master/Casks/w/wezterm.rb)
and have it actually install the terminfo.