add a few missing license files
I didn't symlink to the license file in the root because it mentions
license that don't seem to apply to the crates I added license files
to here.

I used the year each directory was created as the first copyright
year.

I didn't add license files to all crates because I don't need them for
what I'm working on right now, and I want to avoid having to figure
out which licenses apply to which crates.

#2347