macOS: switch to UNUserNotificationCenter for toast notifs
I want to move all the objc stuff over to the objc2 crate.

The older notification center is deprecated and doesn't
have any bindings that are supported by the objc2 crate.

This is mostly straightforward, except I lost hours running
down the delegate was being dropped. Best guess is that
there's some autorelease pool magic that I don't understand;
I tried stashing the object into a global but that still didn't
keep it retained.

In the end, the simplest solution is to take it out of its
Retained wrapper and just leak it; it doesn't hold any
meaningful amount of data anyway, but it is vital to keep
it alive for the notification center callbacks to trigger.

Note that UNUserNotificationCenter requires that the application be
code-signed in order to function. I don't like this, but that's the way
that apple want it: sponsors are supporting the cost of my apple
developer account to enable this.