fix: display whitespace before reset of attributes in InputSelector (#6609)
* fix: display whitespace before reset of attributes in InputSelector

Display white space before reset of background and setting attributes to
default in InputSelector. This is in order for whitespace after the
option to be highlighted in the same background color as the currently
selected option.
The above mentioned points are currently reflected in ShowLauncherArgs
but not in InputSelector.

* fix: emit CRLF after attributes have changed

Moving to the next line before the others are applied might lead to some
wonky behavior in some edge cases. So to be on the safe side, emit CRLF
after attributes have changed